requirements.txt
Flask
stripe
python-dotenv
requests